requirements.txt
azure-functions>=1.18.0
azure-identity>=1.14.0
azure-storage-blob>=12.18.0
python-dateutil>=2.8.2
